[WIP] Fetch data from 10,000ft
This is a very minimal proof of concept for fetching people and project
data from 10,000ft.

Copy `.env.example` to `.env` and put real credentials into it --
there's a utilisation-dashboard user in 1Password that can be used here.

I've started with people rather than with projects for a few reasons:

 - it'll help filter out projects without assignments
 - we want people without projects, so would need to do this anyway
 - we'll want to filter for only billable dxw people (ie not admin-only
   accounts and not tradecraft folk), and this is a better starting
   point for that

This doesn't handle pagination, and uses the default 20-items-per-page,
so will only return 20 out of 70 users. The quickest fix for that would
be to use `users.list(per_page: 100)`, which should be good for the rest
of 2020.

This makes a *lot* of calls to the API: one for all the users, then one
for each user, then one for each project. In total that's probably up to
80 API calls -- given the endpoints I'm not sure that's avoidable, so
expect this script to take a long time to execute and try to make it
resilient to intermittent failures (or find a way to break it down into
multiple shorter scripts).

This currently looks for assignments from yesterday until tomorrow --
given we want to treat team membership as a somewhat more stable thing,
it's probably worth starting by looking at the next 4 weeks. That'll
produce multiple projects for several users, so you'll need to find a
way to work out the primary project.

It's not doing anything to filter out non-project assignments (like
leave), which is (I think!) what results in the projects without names
that get shown.

Finally, this script just prints out the data. It instead needs to use
that data to create entities in the database, so they can be used to
render the dashboard.