Add a column within team members for a tenk_id
10kft has a unique ID for each user. We want to ensure we only create
one TeamMember per user. tenk gives that ID to us as an integer, and
we don't want to allow null values, so we enforce uniqueness.

Both team_member tests have been updated to include a tenk_id as
tenk_id cannot be null.

The uniqueness constraint is not yet implemented.

Our seed file has also been deleted as we are now populating
our database and is not required anymore

Co-authored-by: Olivia <olivia@dxw.com>