Replaced create! with find_or_create_by!
We will be running an update on a set regular time (daily or hourly - yet to be decided)
Once the database is populated the first time, we want to make sure
that any changes will be added to the database.

With a create!, the second time of running the rake task, an error would occur as the data would already exist
within the database and the database cannot add duplicates  (as it can only have unique entries)

The decision to change this to find_or_create_by! means that there will be a check
to see if an entry already exists before adding it.
If it already exisits, it will move onto the next entry.

Co-authored-by: Olivia <olivia@dxw.com>