Test the associations between Projects and TeamMembers
Shows and tests the belong_to relationship of project to a team member and the has_many relationship
of team members to project.

We have left the belong_to relationship which only allows one project for a team member as
we feel this association is enough for our first pass. We will change this to a many to many relationship
at a later date.

Co-authored-by: Olivia <olivia@dxw.com>