Increased the items per page to 100
The default is currently 20 items per page.
This is okay for the short term as we will not be growing as a company beyond 100
team members for the short term.

This will need to be looked at later and pagination to be set up instead.

Changed role to pull discipline data from 10k ft.

Role doesn't best describe the team members role, rather it is their discipline which describe their role