All pull requests require tests to pass before they can be merged
We noticed that tests were not being automatically being run on new pull
requests. This meant that test failures were making it into master. The
immediate problem that caused was that it blocked us from writing new
features.

We started by copying this template work from the Rails template [1]. We then
noticed that builds resulted in a postgres connection issue and that Stuart had
already created a pull request [2] that looked to fix it. Sadly we found
it also produced the same postgres connection issues. Whilst the
postgres service was there, Rails could not establish a connection.

We went through the postgres service documentation [3] to confirm there was
nothing obvious were were doing wrong, without success. We did notice
however that in their example they were making `env` available with
`POSTGRES_HOST` on the step where a connection was expected. A way to
pass the DATABASE_URL to rails is through a single environment variable
of the same name. A refactoring of the database.yml was required for the
testing environment to use this environment variable. Once it did, this
fixed the connection issues. We kept the existing production
configuration so as to not break anything in live environments.

[1]https://github.com/dxw/rails-template
[2]https://github.com/dxw/rails-template/pull/81
[3]https://help.github.com/en/actions/configuring-and-managing-workflows/creating-postgresql-service-containers

Co-authored-by: Tom Hipkin <tomhipkin@gmail.com>