Fix failing tests by making HTTP ENV variables available
We noticed tests were failing and as part of adding CI in to ensure we keep the test suite green we need to go back and test the failing specs.

Autoload dotenv files by using the dotenv-rails [1] rather than the plan ol'Ruby gem.

We could keep on using the dotenv gem and call `Dotenv.load` however we thought we'd stick with the documented steps for Rails apps.

[1]https://github.com/bkeepers/dotenv#rails