Only use the ultimate parent project
For the way we use the team dashboard, we don’t need the different phases of a project.

User assignments in 10kft are done on a project phase basis. Each phase is a project in itself, and the ID of the phase is associated with the user. The phase stores which project it belongs to as a `parent_id` key.

Previously, we were copying the phase - user association as such in our local database, which resulted in the project teams shown on the dashboard being split over project phases.

We are now going up the hierarchy until we find the main project (it will have `parent_id` `nil`), and use the main project to populate the team member assignments in our local database.

This way, the dashboard will show everyone assigned to a project together, regardless of which project phase they are assigned to in 10kft.