Do not store phase name
In the previous commit, we modified the task so that we don’t store different phases of a 10kft project in our local database.

The main (AKA parent) projects do not have a `phase_name`, so this attribute becomes redundant and will be dropped from the table in a future commit.