Test the project fetch rake task
* The project_id column on users no longer looks to be used so there is
no assertion for this field, to avoid clashing when it is removed.
Projects are now retrieved through the new assignment join table.
* As this is a public repo the test data has been anonymised where we
thing appropriate
* We noticed that the test suite is not preventing us from making real
HTTP request, something commonly done with WebMock. As real
authentication variables need to be set in order for the connection to
work we're happy to leave this as a separate problem.

Co-authored-by: Tom <tomh@dxw.com>