Attempt to disambiguate names
The old incarnation of the team dashboard used to have a manual way to disambiguate people with the same first names, in order to help people know who’s who.

This commit adds an automatic way to detect when there are people with the same first names in order to append the initial of their last name, if they have one, to the displayed name.

It doesn’t do anything clever if the last names have the same initial. I have considered this an acceptable edge case to have unresolved.