Sort teams members alphabetically with delivery leads first
Delivery leads are usually the first point of contact when someone wants to find out about a project, so we want to show them first.

In the case of a project with multiple delivery leads, such as GovPress Future, the leads will be sorted by name.

All the other team members will be sorted by name.

Sorting in Rails rather than via database query is OK, because these will always be small numbers of people.