Add email to team member
The email is the only piece of information reliably connecting an OpsGenie user to a 10kft user.

We will need this information in order to assign support rotations (for which the source is OpsGenie) to a team member, so we can have the full picture of what projects a given team member was assigned to be on during their support shift.