Return early if a project cannot be found
A project without an `id` is in fact a project that couldn’t be found, most likely because of some stale data in 10kft. (A user still has an `assignable_id`, but there is no project with that ID anymore.)

There is nothing to do in that case, so we can exit early, and skip the rest of the nil guarding.