Recreate assignments every time
In our local schema, assignments are only a many-to-many glue between teams and projects, and do not have time intervals associated with them.

There is no need to keep historical assignments right now, as the dashboard is meant to give the picture for the current day.

Remove all existing assignments before adding the current ones for the day. This way, older assignments to long-running projects (such as 1st line support) will not accumulate and create confusion.