Add default scope sorted by date
We need to show at least the month before the current week, so that the Head of Tech Ops can check who was on OoH and thus needs extra pay.

We need to show some months in the future, so that the Head of Tech Ops can check how well they're covered for support.

Weeks starting on Wednesday are a dxw-specific quirk. Sprints start on Wednesdays, so support shifts are also aligned to a Wednesday start, to minimise disruption and confusion.

Starting the interval with a Wednesday will produce a table that is best aligned with the dxw working patterns, and thus more intelligible and intuitive within our operating context.