Add rake task to fetch support rotations from API
The latest data from the API (which fetches it from OpsGenie) is always considered the most accurate.

Support assignments can be corrected and overriden, especially for future weeks, as people's schedules may need to change.

We need to wipe out the existing support days, to avoid stale data from showing up on the dashboard.